srdf::srdf

Trait Query

pub trait Query: Rdf {
    // Required method
    fn triples(&self) -> Result<impl Iterator<Item = Self::Triple>, Self::Err>;

    // Provided methods
    fn triples_matching<S, P, O>(
        &self,
        subject: S,
        predicate: P,
        object: O,
    ) -> Result<impl Iterator<Item = Self::Triple>, Self::Err>
       where S: Matcher<Self::Subject>,
             P: Matcher<Self::IRI>,
             O: Matcher<Self::Term> { ... }
    fn triples_with_subject<S: Matcher<Self::Subject>>(
        &self,
        subject: S,
    ) -> Result<impl Iterator<Item = Self::Triple>, Self::Err> { ... }
    fn triples_with_predicate<P: Matcher<Self::IRI>>(
        &self,
        predicate: P,
    ) -> Result<impl Iterator<Item = Self::Triple>, Self::Err> { ... }
    fn triples_with_object<O: Matcher<Self::Term>>(
        &self,
        object: O,
    ) -> Result<impl Iterator<Item = Self::Triple>, Self::Err> { ... }
    fn incoming_arcs(
        &self,
        object: Self::Term,
    ) -> Result<IncomingArcs<Self>, Self::Err> { ... }
    fn outgoing_arcs(
        &self,
        subject: Self::Subject,
    ) -> Result<OutgoingArcs<Self>, Self::Err> { ... }
    fn outgoing_arcs_from_list(
        &self,
        subject: &Self::Subject,
        preds: &[Self::IRI],
    ) -> Result<OutgoingArcsFromList<Self>, Self::Err> { ... }
}

This trait contains functions to handle basic navigation in RDF graphs, with methods that can get triples and the neighbourhood of RDF nodes

Required Methods

fn triples(&self) -> Result<impl Iterator<Item = Self::Triple>, Self::Err>

Provided Methods

fn triples_matching<S, P, O>( &self, subject: S, predicate: P, object: O, ) -> Result<impl Iterator<Item = Self::Triple>, Self::Err>
where S: Matcher<Self::Subject>, P: Matcher<Self::IRI>, O: Matcher<Self::Term>,

Note to implementors: this function needs to retrieve all the triples of the graph. Therefore, for use-cases where the graph is large, this function should be implemented in a way that it does not retrieve all triples at once. As an example, for implementations of SPARQL, this function should be implemented to retrieve just the triples that match the given subject, predicate and object.

fn triples_with_subject<S: Matcher<Self::Subject>>( &self, subject: S, ) -> Result<impl Iterator<Item = Self::Triple>, Self::Err>

fn triples_with_predicate<P: Matcher<Self::IRI>>( &self, predicate: P, ) -> Result<impl Iterator<Item = Self::Triple>, Self::Err>

fn triples_with_object<O: Matcher<Self::Term>>( &self, object: O, ) -> Result<impl Iterator<Item = Self::Triple>, Self::Err>

fn incoming_arcs( &self, object: Self::Term, ) -> Result<IncomingArcs<Self>, Self::Err>

fn outgoing_arcs( &self, subject: Self::Subject, ) -> Result<OutgoingArcs<Self>, Self::Err>

get all outgoing arcs from a subject

fn outgoing_arcs_from_list( &self, subject: &Self::Subject, preds: &[Self::IRI], ) -> Result<OutgoingArcsFromList<Self>, Self::Err>

get outgoing arcs from a node taking into account only a controlled list of preds It returns a HashMap with the outgoing arcs and their values and a list of the predicates that have values and are not in the controlled list.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.