srdf::srdf
Trait SRDF

pub trait SRDF: SRDFBasic {
    // Required methods
    fn predicates_for_subject(
        &self,
        subject: &Self::Subject,
    ) -> Result<HashSet<Self::IRI>, Self::Err>;
    fn objects_for_subject_predicate(
        &self,
        subject: &Self::Subject,
        pred: &Self::IRI,
    ) -> Result<HashSet<Self::Term>, Self::Err>;
    fn subjects_with_predicate_object(
        &self,
        pred: &Self::IRI,
        object: &Self::Term,
    ) -> Result<HashSet<Self::Subject>, Self::Err>;
    fn triples_with_predicate(
        &self,
        pred: &Self::IRI,
    ) -> Result<Vec<Triple<Self>>, Self::Err>;
    fn outgoing_arcs(
        &self,
        subject: &Self::Subject,
    ) -> Result<HasMapOfIriAndItem<Self::IRI, Self::Term>, Self::Err>;
    fn incoming_arcs(
        &self,
        object: &Self::Term,
    ) -> Result<HasMapOfIriAndItem<Self::IRI, Self::Subject>, Self::Err>;
    fn outgoing_arcs_from_list(
        &self,
        subject: &Self::Subject,
        preds: &[Self::IRI],
    ) -> Result<(HasMapOfIriAndItem<Self::IRI, Self::Term>, Vec<Self::IRI>), Self::Err>;

    // Provided method
    fn neighs(
        &self,
        node: &Self::Term,
    ) -> Result<ListOfIriAndTerms<Self::IRI, Self::Term>, Self::Err> { ... }
}

This trait contains functions to handle Simple RDF graphs, which are basically to get the neighbourhood of RDF nodes

TODO: Consider alternative names: RDFGraphOps

Required Methods

fn predicates_for_subject( &self, subject: &Self::Subject, ) -> Result<HashSet<Self::IRI>, Self::Err>

fn objects_for_subject_predicate( &self, subject: &Self::Subject, pred: &Self::IRI, ) -> Result<HashSet<Self::Term>, Self::Err>

fn subjects_with_predicate_object( &self, pred: &Self::IRI, object: &Self::Term, ) -> Result<HashSet<Self::Subject>, Self::Err>

fn triples_with_predicate( &self, pred: &Self::IRI, ) -> Result<Vec<Triple<Self>>, Self::Err>

fn outgoing_arcs( &self, subject: &Self::Subject, ) -> Result<HasMapOfIriAndItem<Self::IRI, Self::Term>, Self::Err>

fn incoming_arcs( &self, object: &Self::Term, ) -> Result<HasMapOfIriAndItem<Self::IRI, Self::Subject>, Self::Err>

fn outgoing_arcs_from_list( &self, subject: &Self::Subject, preds: &[Self::IRI], ) -> Result<(HasMapOfIriAndItem<Self::IRI, Self::Term>, Vec<Self::IRI>), Self::Err>

get outgoing arcs from a node taking into account only a controlled list of preds It returns a HashMap with the outgoing arcs and their values and a list of the predicates that have values and are not in the controlled list.

Provided Methods

fn neighs( &self, node: &Self::Term, ) -> Result<ListOfIriAndTerms<Self::IRI, Self::Term>, Self::Err>

Get the neighbours of a term This code creates an intermediate vector and is not very efficient TODO: return an iterator

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.